requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
opencv-python-headless
numpy
pillow
torch
torchvision
ultralytics
gunicorn
sqlalchemy
psycopg2-binary
